2021/pyproject.toml
[tool.poetry]
name = "2021_aoc"
version = "0.1.0"
description = "2021 Advent of Code"
authors = ["kimv <kimv@sahomeloans.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
black = "^21.11b1"
rich = "^10.15.1"
pyperclip = "^1.8.2"
attrs = "^21.2.0"
numpy = "^1.21.4"
ptpython = "^3.0.20"
networkx = {extras = ["default"], version = "^2.6.3"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
